Implemented UI Headers and Global User View
Implemented persistent and adaptive UI Headers and Global User View.
Toolbar -> Logo constrained next to the fragment label which updates on create in the navview.
Added bypass as guess button on Login page.
Implemented a Global User View which stores the user firebase account information persistently in the app until they sign out of firebase.
Also allows userdata to be passed to each fragment and updated simultaneously and asynchronously.